Organise endGame functions
