+++ tests/freeFunctions/game/endGame/haveMetroidsInvaded.test.cpp
@@ -1,6 +1,6 @@
 #include <vector>
-#include "../../../include/game.hpp"
-#include "../../mockModels/mockMetroid.hpp"
+#include "../../../../include/game.hpp"
+#include "../../../mockModels/mockMetroid.hpp"
 
 using ::testing::NiceMock;
 using ::testing::Return;
